Build towers (#1)
* Adds map zero as a json definition for the map.
This is prep for working on building towers again, to make sure we have
a controlled environment.

* WIP on Build Towers

Added a simple entity to hold cursor information. Perhaps it should be...
not a sprite, but some kind of simple cursor? No, it should be a sprite, its
just that we might want to add some flair to it when rendering it. More on that
later.

* Blagha

* Adds IntentComponent and IntentSystem for coolness.

* More implementation done of the tower stuff

* Enables entering buildmode with a cool offset component

* Refactorings, cleanups and comments

* Stuff

* Intents and stuff works nicely with the tower building functionality.
More dynamic code using anonymous functions here and there.

* So many cool things - mostly docs

* Adds more docs

* Adds new intent build functionality to gamepad input. No time to test,
assume it works for the win