Map editor (#11)
* First commit on the map editor

* WHat the actual fuck is going on?

* Cursor and camera movement working properly. Next step is to add some sections to this bitch

* Adds ability to add and delete sections

* Can save maps in a disgusting unformatted json

* New serialization of maps (not json-based because that sucks)
Enabled zoom in / out in the map editor

* Adds a paint-mode where you walk around and continously paint with the current sectiontype

* Refactored the entire keyboard input for the concept screen so that it uses a finited
state machine and commandmaps to control everything. Soo much fun!

* Updated docs

* Adds keymap description, HUD for the editor that doesn't move and
other legendary features.

* Can add new maps as text files - however, it doesn't load stories,
which is "natural", this needs to be added simply!

* Removed unnecessary project
Added YAML-serialization as an experiment, work in progress due to
the TON of sealed classes I use.

* Can add names of stories to map files - which will do until I have reduced
consequences into pure data that can be executed.
This can happen - sort of like actions for Scene2d. But this will most certainly do
for now!

* Updated docs

* Fixed state machines and exiting the map editor, for bonuses