Refactored and Rewritten AI (#15)
* Refactoring AI stuff, due to bugs

* Enables running the game with just one AI for testing and
evaluating behaviors and such. Ambling looks pretty neat, if you
ask me. Problem is, it consists of ONE giant task, when it should be
a ton of smaller tasks. That's the next step!

* New Leaf Tasks added by the minute.

Now I have some more ones:
FindSection, can take a method and different storage classes
FindPath, finds path from one section to another
Next step - walking that path!

* WIP on fixing leaf tasks but I now feel that the treebuilding
code is not reliable and might need rewriting

* Figured out the issue with componentguard task. Basically, I had mistakenly
implemented it as a Decorator, when it should in fact just be a regular
task. It never worked in any context I was using it and probably never had.

Then I had some control flow issues, which is to be expected with complicated if-
statements and flows, but all in all, it looks pretty excellent now.

And most awesomely of all - my flocking and obstacle avoidance code makes the AI move
smoothly through the map.

* Changed finding random section to make sure we find a section further away from us

* Wow, now it is finally coming together

Some experimenting with the builder syntax to make it more readable.

Its going... OK, I guess?

* Removed an old unused system, more cleanup later

* Fixed build errors