So, I have managed to draw a sprite "skewered" properly, making it, effectively, fake 3D in some way. I still don't really have arms, but I'm very very slowly getting there...
But I also feel that I am re-implementing the entire concept of a 3D game - this is just forcing me to
understand vertices all over again, something I could simply be doing with a 3D game
to begin with. I'll have to think about this for a while.

Perhaps I should implement a 3D game for all of this.

Perhaps.